feat: generic NewsArticle and type NewsEntry
Changed the NewsArticle into generic type with identifier type
specified. The identifier has default type of string.

Changed the NewsService to generic class with identifier type
specified. The identifeir has default type of string.

Introduced a generic NewsEntry to represent the tuple of identifier and
the news article of the identifier.

modified:   lib/NewsService.mjs
- feat: ID generic type for NewsArticle.
- feat: ID generic type for NewsService
- feat: type NewsEntry<ID> for tuple containing identifier-article pair.
- refactor: Replace entry tuples with NewsEntry